Format and move FileDownloadsTable.tsx to files
